+++ lib/shared/src/mentions/api.ts
@@ -1,8 +1,9 @@
 import type { ContextItem, ContextItemWithContent } from '../codebase-context/messages'
 import type { Configuration } from '../configuration'
+import { openCtx } from '../context/openctx/api'
+import { logDebug } from '../logger'
 import type { PromptString } from '../prompt/prompt-string'
 import { GITHUB_CONTEXT_MENTION_PROVIDER } from './providers/githubMentions'
-import { OPENCTX_CONTEXT_MENTION_PROVIDER } from './providers/openctxMentions'
 import { PACKAGE_CONTEXT_MENTION_PROVIDER } from './providers/packageMentions'
 import { SOURCEGRAPH_SEARCH_CONTEXT_MENTION_PROVIDER } from './providers/sourcegraphSearch'
 import { URL_CONTEXT_MENTION_PROVIDER } from './providers/urlMentions'
@@ -17,13 +18,11 @@ export type ContextMentionProviderID = string
  *
  * This API is *experimental* and subject to rapid, unannounced change.
  *
- * In VS Code, use {@link getEnabledContextMentionProviders} instead of this.
  */
 export const CONTEXT_MENTION_PROVIDERS: ContextMentionProvider[] = [
     URL_CONTEXT_MENTION_PROVIDER,
     PACKAGE_CONTEXT_MENTION_PROVIDER,
     SOURCEGRAPH_SEARCH_CONTEXT_MENTION_PROVIDER,
-    OPENCTX_CONTEXT_MENTION_PROVIDER,
     GITHUB_CONTEXT_MENTION_PROVIDER,
 ]
 
@@ -111,16 +110,37 @@ export const SYMBOL_CONTEXT_MENTION_PROVIDER: ContextMentionProviderMetadata<'sy
 }
 
 /** Metadata for all registered {@link ContextMentionProvider}s. */
-export function allMentionProvidersMetadata(
+export async function allMentionProvidersMetadata(
     config: Pick<Configuration, 'experimentalNoodle' | 'experimentalURLContext'>
-): ContextMentionProviderMetadata[] {
-    return [
+): Promise<ContextMentionProviderMetadata[]> {
+    const items = [
         FILE_CONTEXT_MENTION_PROVIDER,
         SYMBOL_CONTEXT_MENTION_PROVIDER,
-        ...CONTEXT_MENTION_PROVIDERS.filter(
-            ({ id }) =>
-                config.experimentalNoodle ||
-                (id === URL_CONTEXT_MENTION_PROVIDER.id && config.experimentalURLContext)
-        ),
+        ...(await openCtxMentionProviders()),
     ]
+
+    return items
+}
+
+export async function openCtxMentionProviders(): Promise<ContextMentionProviderMetadata[]> {
+    const client = openCtx.client
+    if (!client) {
+        return []
+    }
+
+    try {
+        const providers = await client.meta({})
+
+        return providers
+            .filter(provider => provider.features?.mentions)
+            .map(provider => ({
+                id: provider.providerUri,
+                title: provider.name + ' (by OpenCtx)',
+                queryLabel: `Search using ${provider.name} provider`,
+                emptyLabel: 'No results found',
+            }))
+    } catch (error) {
+        logDebug('openctx', `Failed to fetch OpenCtx providers: ${error}`)
+        return []
+    }
 }